// src/types/Experience.ts
export interface Experience {
    title: string;          // Job title
    company: string;        // Company name
    location: string;       // Location of the job
    period: string;         // Duration of employment
    description: string[];  // List of responsibilities/accomplishments
    technologies: string[]; // List of technologies used
}